add sin, pow doesn't work

+++ s21_ceil.c
@@ -1,7 +1,7 @@
 #include "s21_math.h"
 
 long double s21_ceil(double x) {
-    int result = 0;
+    double result = 0;
     int check_minus = 0;
 
     if (x < 0) {
@@ -14,5 +14,5 @@ long double s21_ceil(double x) {
     }
     check_minus == 1 ? result-- : 0;
     check_minus == 1 ? result = -result : 0;
-    return result;
+    return (long double) result;
 } // возвращает ближайшее целое число, не меньшее заданного значения
